// src/modules/types.d.ts
type Filter = {
  value: boolean;
  title: string;
};

type FiltersState = {
  all: Filter;
  noTransfers: Filter;
  oneTransfer: Filter;
  twoTransfers: Filter;
  threeTransfers: Filter;
};

type Ticket = {
  // Цена в рублях
  price: number;
  // Код авиакомпании (iata)
  carrier: string;
  // Массив перелётов.
  // В тестовом задании это всегда поиск "туда-обратно" значит состоит из двух элементов
  segments: [
    {
      // Код города (iata)
      origin: string;
      // Код города (iata)
      destination: string;
      // Дата и время вылета туда
      date: string;
      // Массив кодов (iata) городов с пересадками
      stops: string[];
      // Общее время перелёта в минутах
      duration: number;
    },
    {
      // Код города (iata)
      origin: string;
      // Код города (iata)
      destination: string;
      // Дата и время вылета обратно
      date: string;
      // Массив кодов (iata) городов с пересадками
      stops: string[];
      // Общее время перелёта в минутах
      duration: number;
    },
  ];
};

type TicketId = string;

type TicketsState = {
  searchId: string | null;
  status: 'loading' | 'rejected' | 'resolved';
  ticketsData: { stop: boolean; tickets: Ticket[] };
  tickets: {
    entities: Record<TicketId, Ticket> | object;
    ids: TicketId[];
  };
};

type SortState = {
  sortValue: 'cheapest' | 'fastest' | 'optimal';
};

type TicketsShownState = number;

export { Filter, FiltersState, TicketsShownState, SortState, Ticket, TicketId, TicketsState };